Add cast overload for option in various types.
Remove double cast.
Add public getter for min and max value of an option to make them accessible.

From the UCI specifications:

type The option has type t. There are 5 different types of options the engine can send

- check a checkbox that can either be true or false
- spin a spin wheel that can be an integer in a certain range
- combo a combo box that can have different predefined strings as a value
- button a button that can be pressed to send a command to the engine
- string a text field that has a string as a value, an empty string has the value ""

This commit intents to make the StockFish code compliant to the UCI definition for handling UCI types for options.

+++ src/search.cpp
@@ -269,7 +269,7 @@ void MainThread::search() {
   Thread* bestThread = this;
 
   // Check if there are threads with a better score than main thread
-  if (    Options["MultiPV"] == 1
+  if ( size_t(Options["MultiPV"]) == 1
       && !Limits.depth
       && !(Skill(Options["Skill Level"]).enabled() || Options["UCI_LimitStrength"])
       &&  rootMoves[0].pv[0] != MOVE_NONE)
@@ -350,8 +350,9 @@ void Thread::search() {
   // for match (TC 60+0.6) results spanning a wide range of k values.
   PRNG rng(now());
   double floatLevel = Options["UCI_LimitStrength"] ?
-                        clamp(std::pow((Options["UCI_Elo"] - 1346.6) / 143.4, 1 / 0.806), 0.0, 20.0) :
-                        double(Options["Skill Level"]);
+    clamp(std::pow(( int( Options["UCI_Elo"] ) - 1346.6) / 143.4, 1 / 0.806), 0.0, 20.0) :
+    int(Options["Skill Level"]);
+
   int intLevel = int(floatLevel) +
                  ((floatLevel - int(floatLevel)) * 1024 > rng.rand<unsigned>() % 1024  ? 1 : 0);
   Skill skill(intLevel);
